+++ mongo/document/base.go
@@ -5,6 +5,8 @@ import (
 	"time"
 
 	"gopkg.in/mgo.v2/bson"
+
+	"github.com/Scalingo/go-utils/errors/v2"
 )
 
 type Base struct {
@@ -45,6 +47,6 @@ func (d *Base) destroy(ctx context.Context, collection string) error {
 	return ReallyDestroy(ctx, collection, d)
 }
 
-func (d *Base) Validate(_ context.Context) (*ValidationErrors, error) {
+func (d *Base) Validate(_ context.Context) (*errors.ValidationErrors, error) {
 	return nil, nil
 }